requirements.txt
django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
pytest==8.3.2
pytest-django==4.9.0
factory_boy==3.3.0
debugpy==1.8.5